move common styles to components folder
